// packages/contract/src/generated/SessionsUpdatedEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PublicReferencedSessionHead } from "./PublicReferencedSessionHead";
import type { SessionReference } from "./SessionReference";
import type { SessionsUpdatedEventType } from "./SessionsUpdatedEventType";

export type SessionsUpdatedEvent = {
  type: SessionsUpdatedEventType;
  changedAgents: Array<string>;
  newSessionRefs: Array<SessionReference>;
  totalSessions: number;
  timestamp: number;
  changedSessionHeads: Array<PublicReferencedSessionHead>;
  projectionRelatedSessionHeads?: Array<PublicReferencedSessionHead>;
  projectionSessionOrder?: Array<SessionReference>;
  removedSessionRefs: Array<SessionReference>;
};